grpclb: rename `grpclbstate` package back to `state` (#5962)
Fixes https://github.com/grpc/grpc-go/issues/5928
